feat: replace code to be numeric, a bit of tidy up
