--- src/screens/Create.tsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import { View, Text } from 'react-native';
-
-const Create = () => (
-  <View>
-    <Text>create</Text>
-  </View>
-);
-
-export default Create;
