+++ src/screens/Drawer/ShareTheApp.tsx
@@ -2,7 +2,7 @@ import React, { FC } from 'react';
 import { View, StyleSheet, Image } from 'react-native';
 import { useDispatch, useSelector } from 'react-redux';
 import * as Analytics from 'appcenter-analytics';
-import Share, { Options } from 'react-native-share';
+import Share from 'react-native-share';
 import { Container, PrimaryButton, AppText } from '../../common';
 import assets from '../../assets';
 import { getPlatformDimension } from '../../utils/device';
@@ -10,21 +10,14 @@ import { analytics } from '../../utils';
 import { sentInviteToFriends } from '../../redux/user/actions';
 import { RootState } from '../../redux/store';
 import { addFreeDeck } from '../../redux/decks/actions';
-
-const options: Options = {
-  url: 'https://myflashcards.app',
-  message: 'Check out a new app called MyFlashCards',
-  title: '',
-  subject: 'Learn with Flashcards App',
-  saveToFiles: false,
-};
+import { shareOptions } from '../../config';
 
 const ShareTheApp: FC = () => {
   const dispatch = useDispatch();
   const { user } = useSelector((state: RootState) => state);
   const handlePressInvite = () => {
     Analytics.trackEvent(analytics.inviteFriends).catch(null);
-    return Share.open(options)
+    return Share.open(shareOptions)
       .then(() => {
         if (!user.hasSentInvite) {
           dispatch(sentInviteToFriends());
@@ -40,7 +33,7 @@ const ShareTheApp: FC = () => {
       <View style={styles.imageContainer}>
         <Image source={assets.icons.faces} resizeMode="contain" />
       </View>
-      <View style={{ flex: 1 }}>
+      <View style={styles.content}>
         <AppText centered size="h2">
           Enjoying the app?
         </AppText>
@@ -52,9 +45,7 @@ const ShareTheApp: FC = () => {
             & get extra free deck!
           </AppText>
         )}
-        <View style={styles.buttonContainer}>
-          <PrimaryButton buttonText="Share" onPress={handlePressInvite} />
-        </View>
+        <PrimaryButton buttonText="Share" onPress={handlePressInvite} buttonStyle={styles.buttonContainer} />
       </View>
     </Container>
   );
@@ -66,6 +57,9 @@ const styles = StyleSheet.create({
     alignSelf: 'center',
     alignContent: 'center',
     backgroundColor: '#fff',
+  },
+  content: {
+    flex: 1,
   },
   imageContainer: {
     flex: 1.6,
