+++ src/screens/Home/AddDeck/index.tsx
@@ -4,7 +4,7 @@ import { StackNavigationProp } from '@react-navigation/stack';
 import { View, Animated, StyleSheet } from 'react-native';
 import { CloseButton } from '../../../common';
 import { saveDeck } from '../../../redux/decks/actions';
-import { RootStackParamList, Screens } from '../../../navigation/interface';
+import { RootStackParamList, Screens } from '../../../navigation/types';
 import { getPlatformDimension, isIOS } from '../../../utils/device';
 import { selectAllDecks, selectMaxFreeDecks } from '../../../redux/seclectors';
 import { AddDeckContent } from './AddDeckContent';
